Add bootloader pattern for all roles
YaST will install all the packages that are resolved from the patterns
listed in the control.xml (for a selected role).  Internally also have
a list of packages that are installed that do not come from any
external requirement, for example the kernel or the Grub bootloader.

Since some time ago there is a pattern bootloader in the pattern-base
package, that can be used to explicitly drive an installation that
requires a bootloader.  This is an exercise to re-use the same
patterns for containers, that will not require a bootloader.

This patch makes explicit the requirement of the bootloader pattern.